Survo R 0.6.3 - Specifications FILTER=#DOTCOMMA and FILTER=#COMMADOT can be used to replace dots with commas (and vice versa) in FILE SAVE - Autoconversion of single and double quotations to ASCII when coppying from clipboard - If default Tcl/Tk clipboard fetching returns a zero length STRING, contents are fetched again using type UTF8_STRING - Changed version check in /LM-ANOVA sucro - Check for mask error in FILE LOAD <data> TO R>data - Notification if MATCH-fields are of different type in FILE COPY - Some extra printing to screen (by output_line as in STAT or TAB) can be skipped by using specification PRIND=0 - Fixes for path and output_open checks in COPY L1,L2 TO <file> - Fix for maskline parsing while copying data matrix from edit field to data file - Fix for partially overwritten masks while opening data matrix in edit field with maskline - Path to start sucro can now be given with command survo(sucro="<Path to start sucro>")) - Path to config file (such as MUSTE.APU) can be given with parameter config: survo(config="<Path to start config file">) - In FILE SAVE, FIELDS-list is now by default searched only until next borderline - FIELDS=<name> can be used in FILE SAVE to indicate fields-list to be used - Updated START.EDT and DEMOS.EDT - NEWS text file included

+++ pkg/muste/src/data.c
@@ -1336,7 +1336,7 @@ int ma_open(char *name,SURVO_DATA_MATRIX *s,int drivi)
         s->vartype=(char **)p; p+=m*sizeof(char *);
         for (i=0; i<m; ++i) s->vartype[i]=tyypit+(TYPELEN+1)*i;
         s->varlen=(short *)p; p+=m*sizeof(short);
-        s->varpos=(int *)p; p+=m*sizeof(short); // RS 28.1.2013 varpos short -> int
+        s->varpos=(int *)p; p+=m*sizeof(int); // RS 28.1.2013 varpos short -> int
         masktila=p; p+=LLENGTH;
         s->mask=(char **)p; p+=m*sizeof(char *);
         s->obs=p; p+=LLENGTH;
@@ -1374,8 +1374,8 @@ int ma_open(char *name,SURVO_DATA_MATRIX *s,int drivi)
         else
             {
             lmask=edline2(dsana[5],1,1); if (lmask==0) return(-1);
-            edread(masktila,lmask);
-            i=split(masktila+1,s->mask,m);
+            edread(masktila,lmask);             
+            i=split(masktila+1,s->mask,m);                    
             if (i<m)
                 {
                 sprintf(sbuf,"\nIncorrect mask line for DATA %s",name); sur_print(sbuf);
@@ -1384,13 +1384,13 @@ int ma_open(char *name,SURVO_DATA_MATRIX *s,int drivi)
                 }
 
             for (i=0; i<m; ++i)
-                {
+                {                                    
                 s->varpos[i]=s->mask[i]-masktila;
-                s->varlen[i]=strlen(s->mask[i]);
-                s->vartype[i][1]=s->mask[i][0];
-                }
-            }
-        *active_data=EOS;
+                s->varlen[i]=strlen(s->mask[i]);              
+                s->vartype[i][1]=s->mask[i][0];              
+                }               
+            }                   
+        *active_data=EOS;     
         return(1);
         }
 
@@ -2334,11 +2334,12 @@ int data_save(SURVO_DATA *d, long j, int i, double x)
 
         varlen=d->varlen[i];
         strncpy(sana,space,(unsigned int)varlen); sana[varlen-1]='-';
+// Rprintf("\nmask: %s",d->d1.mask[i]);         
         if (fabs(x)<MISSING8/1000.0)
             {
             k=fconv(x,d->d1.mask[i],sana2);
             if (k>=0) strcpy(sana,sana2);
-            }
+            }           
         ma_save(&(d->d1),(int)j,i,sana);
         return(1);
         }
